Update libopenhmd.so to experimental Oculus tracking.
This uses the libopenhmd.so which is a mildly hacked version - only disabling some
annoying debug messages, of the current experimental rift-kalman-filter branch from
https://github.com/thaytan/OpenHMD.git

This version has experimental support for using the Oculus CV1 sensors to
track the position and orientation of the Oculus Rift CV1 and the 1st gen
Oculus touch controllers. It also has support for linear and angular velocity
and acceleration reporting, and for enabling haptic feedback on the Oculus
touch controllers.

This is not yet merged upstream in OpenHMD public repo, but it is working
well enough for all my needs on Linux atm., so others should be able to
benefit from it as well.